Major sync up of SICD/SIDD 1.0 support
git-svn-id: https://wush.net/svn/mdaus-oss/trunk@356 9d772353-a6c9-4ee9-b2cc-df47a7649cd5

+++ modules/c++/samples/test_parse_xml.cpp
@@ -364,9 +364,9 @@ void generateKMLForSICD(xml::lite::Element* docXML,
     // Time associated with start
     six::DateTime dateTime = data->timeline->collectStart;
 
-    int durationInSeconds = (int) data->timeline->collectDuration;
+    size_t durationInSeconds = (size_t) data->timeline->collectDuration;
     six::Vector3 atTimeX;
-    for (unsigned int i = 0; i < durationInSeconds; ++i)
+    for (size_t i = 0; i < durationInSeconds; ++i)
     {
         atTimeX = data->position->arpPoly(i);
         v[1] = scene::Utilities::ecefToLatLon(atTimeX);
