+++ modules/c++/six.sidd/include/six/sidd/GeoTIFFReadControl.h
@@ -1,10 +1,10 @@
 /* =========================================================================
- * This file is part of six-c++ 
+ * This file is part of six.sidd-c++ 
  * =========================================================================
  * 
  * (C) Copyright 2004 - 2009, General Dynamics - Advanced Information Systems
  *
- * six-c++ is free software; you can redistribute it and/or modify
+ * six.sidd-c++ is free software; you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
  * the Free Software Foundation; either version 3 of the License, or
  * (at your option) any later version.
@@ -51,7 +51,7 @@ public:
 
     virtual void load(const std::string& fromFile);
 
-    virtual UByte* interleaved(Region& region, int imageNumber);
+    virtual UByte* interleaved(Region& region, size_t imageNumber);
 
     virtual std::string getFileType() const
     {
